Run prettier formatting on all files and create components for certain UI to make code more readable
